Contact List (Express.js)
This version is not yet connected to the database. In this contact-list version I have demonstrated addition and deletion of contacts. 
Practical implementation of how browser sends a request via router(URL) to the controller and the controller either update the form or data in HTML file (view) and sends it back to the browser.